Update scaling fixes in utils

+++ lib/addm.cpp
@@ -641,9 +641,9 @@ vector<aDDMTrial> aDDMTrial::loadTrialsFromCSV(string filename) {
         std::getline(ss, field, ',');
         RT = std::stoi(field);
         std::getline(ss, field, ',');
-        valueLeft = std::stoi(field);
+        valueLeft = std::stof(field);
         std::getline(ss, field, ',');
-        valueRight = std::stoi(field);
+        valueRight = std::stof(field);
         std::getline(ss, field, ',');
         fItem = std::stoi(field);
         std::getline(ss, field, ',');
